Refactor project structure and improve core components
Reorganized file hierarchy for better modularity and clarity, including moving `Camera` and `Window` to `graphics` and `window` directories, respectively. Introduced smart pointers for resource management and replaced `float` with `double` precision in multiple areas. Added `Renderer` and `UIManager` stubs for future development.

+++ src/graphics/Camera.cpp
@@ -1,19 +1,19 @@
 #include "Camera.h"
 
-Camera::Camera(glm::vec3 startPosition)
-    : position(startPosition), front(glm::vec3(0.0f, 0.0f, -1.0f)), up(glm::vec3(0.0f, 1.0f, 0.0f)),
-      yaw(-90.0f), pitch(0.0f), speed(2.5f), sensitivity(0.1f) {}
+Camera::Camera(glm::dvec3 startPosition)
+    : position(startPosition), front(glm::dvec3(0.0, 0.0, -1.0)), up(glm::dvec3(0.0, 1.0, 0.0)),
+      yaw(-90.0), pitch(0.0), speed(2.5), sensitivity(0.1) {}
 
-glm::mat4 Camera::getViewMatrix() {
+glm::dmat4 Camera::getViewMatrix() {
     return glm::lookAt(position, position + front, up);
 }
 
-void Camera::move(const glm::vec3& direction, float deltaTime) {
-    float velocity = speed * deltaTime;
+void Camera::move(const glm::dvec3& direction, double deltaTime) {
+    double velocity = speed * deltaTime;
     position += direction * velocity;
 }
 
-void Camera::rotate(float xOffset, float yOffset) {
+void Camera::rotate(double xOffset, double yOffset) {
     xOffset *= sensitivity;
     yOffset *= sensitivity;
 
@@ -23,7 +23,7 @@ void Camera::rotate(float xOffset, float yOffset) {
     if (pitch > 89.0f) pitch = 89.0f;
     if (pitch < -89.0f) pitch = -89.0f;
 
-    glm::vec3 direction;
+    glm::dvec3 direction;
     direction.x = cos(glm::radians(yaw)) * cos(glm::radians(pitch));
     direction.y = sin(glm::radians(pitch));
     direction.z = sin(glm::radians(yaw)) * cos(glm::radians(pitch));
